fix: expense submission 500 error - rule_id column and error handling

+++ backend/models/Expense.js
@@ -64,7 +64,7 @@ async function updateStatus(id, status) {
 async function setRuleId(id, ruleId) {
     // Add rule_id column if it doesn't exist, then store the value
     await pool.query(`
-        ALTER TABLE expenses ADD COLUMN IF NOT EXISTS rule_id INTEGER REFERENCES approval_rules(id)
+        ALTER TABLE expenses ADD COLUMN IF NOT EXISTS rule_id BIGINT REFERENCES approval_rules(id)
     `);
     const query = `UPDATE expenses SET rule_id = $2 WHERE id = $1 RETURNING *`;
     const { rows } = await pool.query(query, [id, ruleId]);
